--- Assets/01. Scenes/CDG.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e74d5e7dd8f23a04c99ed09e3411733f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
