--- Assets/01. Scenes/CDG/Codes(UI).meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dbed4986c74ef8b4abb5b9c334ed3b95
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
