--- Assets/01. Scenes/CDG/Codes(UI)/DialogueManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 97ad5236a12ccde4f95fe68e0a1e399a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
