--- Assets/01. Scenes/CDG/StatusBoxAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b7c2800501bf160498a4ef49558c7f29
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
