--- Assets/01. Scenes/CDG/StatusBoxAssets/HPBar.anim
@@ -1,99 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!74 &7400000
-AnimationClip:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: HPBar
-  serializedVersion: 7
-  m_Legacy: 0
-  m_Compressed: 0
-  m_UseHighQualityCurve: 1
-  m_RotationCurves: []
-  m_CompressedRotationCurves: []
-  m_EulerCurves: []
-  m_PositionCurves: []
-  m_ScaleCurves: []
-  m_FloatCurves: []
-  m_PPtrCurves:
-  - serializedVersion: 2
-    curve:
-    - time: 0
-      value: {fileID: -1517190800220980965, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.083333336
-      value: {fileID: -179325417656630464, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.16666667
-      value: {fileID: -6613570599918036523, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.25
-      value: {fileID: 3464853799452460196, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.33333334
-      value: {fileID: -5343482278918804089, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.41666666
-      value: {fileID: 5776891871705209354, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.5
-      value: {fileID: -7841385830542087195, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.5833333
-      value: {fileID: 2569998892846186871, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.6666667
-      value: {fileID: -3633511616331905624, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - time: 0.75
-      value: {fileID: 7411263617574662519, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    attribute: m_Sprite
-    path: 
-    classID: 114
-    script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
-    flags: 2
-  m_SampleRate: 60
-  m_WrapMode: 0
-  m_Bounds:
-    m_Center: {x: 0, y: 0, z: 0}
-    m_Extent: {x: 0, y: 0, z: 0}
-  m_ClipBindingConstant:
-    genericBindings:
-    - serializedVersion: 2
-      path: 0
-      attribute: 2015549526
-      script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
-      typeID: 114
-      customType: 0
-      isPPtrCurve: 1
-      isIntCurve: 0
-      isSerializeReferenceCurve: 0
-    pptrCurveMapping:
-    - {fileID: -1517190800220980965, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: -179325417656630464, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: -6613570599918036523, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: 3464853799452460196, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: -5343482278918804089, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: 5776891871705209354, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: -7841385830542087195, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: 2569998892846186871, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: -3633511616331905624, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-    - {fileID: 7411263617574662519, guid: ad304d4b2b349814cba8c8ec437a2599, type: 3}
-  m_AnimationClipSettings:
-    serializedVersion: 2
-    m_AdditiveReferencePoseClip: {fileID: 0}
-    m_AdditiveReferencePoseTime: 0
-    m_StartTime: 0
-    m_StopTime: 0.76666665
-    m_OrientationOffsetY: 0
-    m_Level: 0
-    m_CycleOffset: 0
-    m_HasAdditiveReferencePose: 0
-    m_LoopTime: 0
-    m_LoopBlend: 0
-    m_LoopBlendOrientation: 0
-    m_LoopBlendPositionY: 0
-    m_LoopBlendPositionXZ: 0
-    m_KeepOriginalOrientation: 0
-    m_KeepOriginalPositionY: 1
-    m_KeepOriginalPositionXZ: 0
-    m_HeightFromFeet: 0
-    m_Mirror: 0
-  m_EditorCurves: []
-  m_EulerEditorCurves: []
-  m_HasGenericRootTransform: 0
-  m_HasMotionFloatCurves: 0
-  m_Events: []
