--- Assets/01. Scenes/CDG/StatusBoxAssets/HPBar.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 50ad70f09c5b6b74eb6dfcb427bf42e7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
