--- Assets/01. Scenes/CDG/StatusBoxAssets/HPBar.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9f62c59284914d0499ad7c4b417e7ac6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
