--- Assets/01. Scenes/CDG/StatusBoxAssets/StatusPanel.aseprite.meta
@@ -1,164 +1,0 @@
-fileFormatVersion: 2
-guid: 6fd027d91d259cb468b0a041b240e548
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 62a9f0aa5b59740cfbadc7e5f9823bb0, type: 3}
-  textureImporterSettings:
-    alphaSource: 1
-    mipMapMode: 0
-    enableMipMap: 0
-    fadeOut: 0
-    borderMipMap: 0
-    mipMapsPreserveCoverage: 0
-    alphaTestReferenceValue: 0.5
-    mipMapFadeDistanceStart: 1
-    mipMapFadeDistanceEnd: 3
-    convertToNormalMap: 0
-    heightScale: 0.25
-    normalMapFilter: 0
-    flipGreenChannel: 0
-    swizzle: 50462976
-    isReadable: 0
-    streamingMipmaps: 0
-    streamingMipmapsPriority: 0
-    vTOnly: 0
-    ignoreMipmapLimit: 0
-    nPOTScale: 1
-    sRGBTexture: 1
-    spriteMode: 2
-    spriteExtrude: 1
-    spriteMeshType: 1
-    alignment: 0
-    spritePivot: {x: 0.5, y: 0.5}
-    spritePixelsToUnits: 100
-    spriteBorder: {x: 0, y: 0, z: 0, w: 0}
-    spriteGenerateFallbackPhysicsShape: 0
-    generateCubemap: 6
-    cubemapConvolution: 0
-    seamlessCubemap: 0
-    alphaIsTransparency: 1
-    spriteTessellationDetail: -1
-    textureType: 8
-    textureShape: 1
-    singleChannelComponent: 0
-    flipbookRows: 0
-    flipbookColumns: 0
-    ignorePngGamma: 0
-    cookieMode: 0
-    filterMode: 0
-    aniso: 1
-    mipBias: 0
-    wrapU: 1
-    wrapV: 1
-    wrapW: 1
-    normalMap: 0
-    textureFormat: 0
-    maxTextureSize: 0
-    lightmap: 0
-    compressionQuality: 0
-    linearTexture: 0
-    grayScaleToAlpha: 0
-    rGBM: 0
-    cubemapConvolutionSteps: 0
-    cubemapConvolutionExponent: 0
-    maxTextureSizeSet: 0
-    compressionQualitySet: 0
-    textureFormatSet: 0
-    applyGammaDecoding: 0
-  previousAsepriteImporterSettings:
-    fileImportMode: 1
-    importHiddenLayers: 0
-    layerImportMode: 1
-    defaultPivotSpace: 0
-    defaultPivotAlignment: 7
-    customPivotPosition: {x: 0.5, y: 0.5}
-    spritePadding: 0
-    generateModelPrefab: 1
-    generateAnimationClips: 1
-    addSortingGroup: 1
-    addShadowCasters: 0
-  asepriteImporterSettings:
-    fileImportMode: 1
-    importHiddenLayers: 0
-    layerImportMode: 1
-    defaultPivotSpace: 0
-    defaultPivotAlignment: 7
-    customPivotPosition: {x: 0.5, y: 0.5}
-    spritePadding: 0
-    generateModelPrefab: 1
-    generateAnimationClips: 1
-    addSortingGroup: 1
-    addShadowCasters: 0
-  importFileNodeState: 1
-  platformSettingsDirtyTick: 0
-  textureAssetName: 
-  singleSpriteImportData:
-  - name: 
-    originalName: 
-    pivot: {x: 0, y: 0}
-    alignment: 0
-    border: {x: 0, y: 0, z: 0, w: 0}
-    rect:
-      serializedVersion: 2
-      x: 0
-      y: 0
-      width: 0
-      height: 0
-    spriteID: 
-    spriteBone: []
-    spriteOutline: []
-    vertices: []
-    spritePhysicsOutline: []
-    indices: 
-    edges: []
-    tessellationDetail: 0
-    uvTransform: {x: 0, y: 0}
-  animatedSpriteImportData:
-  - name: StatusPanel
-    originalName: 
-    pivot: {x: 0.5, y: 0}
-    alignment: 9
-    border: {x: 0, y: 0, z: 0, w: 0}
-    rect:
-      serializedVersion: 2
-      x: 4
-      y: 4
-      width: 32
-      height: 32
-    spriteID: a8c424a8a40fea94ea8cbf61440ebb72
-    spriteBone: []
-    spriteOutline: []
-    vertices: []
-    spritePhysicsOutline: []
-    indices: 
-    edges: []
-    tessellationDetail: 0
-    uvTransform: {x: 4, y: 4}
-  spriteSheetImportData: []
-  asepriteLayers:
-  - layerIndex: 0
-    guid: -1859550654
-    name: StatusPanel
-    layerFlags: 0
-    layerType: 0
-    blendMode: 0
-    cells:
-    - name: StatusPanel
-      frameIndex: 0
-      cellRect:
-        x: 0
-        y: 0
-        width: 32
-        height: 32
-      spriteId: a8c424a8a40fea94ea8cbf61440ebb72
-    linkedCells: []
-    parentIndex: -1
-  platformSettings: []
-  secondarySpriteTextures: []
-  spritePackingTag: 
-  canvasSize: {x: 32, y: 32}
