--- Assets/01. Scenes/CDG/TextBoxAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0fa3e1c4055c2ce45baf3324788ab6ee
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
