--- Assets/01. Scenes/CDG/TextBoxAssets/WaitCursor-Animation.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6aa631cc135e98b4090d449ce95f1575
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
