--- Assets/01. Scenes/CDG/TextBoxAssets/WaitCursor-AnimationController.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b02ce2221fac5548a90ca3a4e5d573d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
