--- Assets/01. Scenes/CDG/UI(TextBox).unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ecdf178163bf75a4ebc6d724290c7464
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
