--- Assets/01. Scenes/OMH.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 64b161e4dfa692c488ae5877a2a73464
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
