--- Assets/01. Scenes/OMH/Inventory.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c3af34ff5a4b33c46867b08178b57c4e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
