--- Assets/01. Scenes/OMH/InventoryUI.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9b218025bae496b47a13a5949687f724
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
