--- Assets/01. Scenes/OMH/Items.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f86dc54d0fbe4d947aee36715070581e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
