--- Assets/01. Scenes/OMH/Items/Items.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 37f5663facfcd6c4c97bffc0a5694fbc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
