+++ Assets/01. Scenes/StoryScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e604bd86756994844883bd5e7fdc89d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
