+++ Assets/01. Scenes/StoryScene/Inventory.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c3af34ff5a4b33c46867b08178b57c4e
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
