+++ Assets/01. Scenes/StoryScene/InventoryUI.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9b218025bae496b47a13a5949687f724
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
