+++ Assets/01. Scenes/StoryScene/StoryInterface.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ecdf178163bf75a4ebc6d724290c7464
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
