+++ Assets/02. Scripts/UI/Choice.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6b89d0b16aff7a479a594c331b8dd6e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
