+++ Assets/Resources/StoryScript.txt
@@ -1,6 +1,5 @@
-김똘똘#테스트 파일 읽어 오기 - 1
-김똘순#테스트 파일 읽어 오기 - 2
-김똘똘#테스트 파일 읽어 오기 - 3
-김똘순#테스트 파일 읽어 오기 - 4
-김똘똘#0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000
+여주#테스트 파일 읽어 오기 - 1
+좀비#테스트 파일 읽어 오기 - 2
+여주#테스트 파일 읽어 오기 - 3
+좀비#테스트 파일 읽어 오기 - 4
 *#*
