+++ Packages/manifest.json
@@ -1,6 +1,9 @@
 {
   "dependencies": {
+    "com.unity.2d.aseprite": "1.0.1",
     "com.unity.collab-proxy": "2.3.1",
+    "com.unity.ide.visualstudio": "2.0.22",
+    "com.unity.ide.vscode": "1.2.5",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.7.5",
     "com.unity.toolchain.win-x86_64-linux-x86_64": "2.0.6",
